Issue 42 unary expressions (#43)
* Start of UnaryOpNodes.

* Updated parsing logic for parsing unary op nodes.

Updated parsing for testing with FunctionCallNodes.

* Refactoring previous changes for clarity.  AddressOfNode is now
deprecated, use UnaryOpNode instead.  Keeping AddressOfNode class for
now in case we decide to re-use in the future.  This ofcourse will cause
warnings for deprecated code use (ignoring because of this).

* Re-implement AddressOfNode analysis as a special case in UnaryOpNode
analysis.  Added relevant unittests as well.

* Added increment and decrement parse tests along with a few analyzer test cases.

Fixed a unit test and discovered a new issue with the type system...

* Resolve review comments.