Feat (Encode): Codegen for IfElseStatementNodes and related changes (WIP).
This is much more involved than I anticipated.  Essentially, this will
require the Analyzer to be updated to tweak the AST to always
ensure the ifConditionalExpression is a comparator.  Which,
means we now need to support these additional EnumBinOpTypes...

Here is a brief overview of what was updated/changed:
- This required adding logic to wrap a if-conditional expression in the
Analyzer.
- Additional work was done to partial implement comparision operators
(excluding the Parser).

TODO (at least what I could come up with for this commit):
- Update Parser to support the remaining comparison operators.
- Update Analyzer to support remaining comparision operators.
- Update Encode(r) to support remaining comparision operators.